requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.2+cu118
torchaudio>=2.1.2+cu118
torchvision>=0.16.2+cu118
annotated_types==0.6.0
coloredlogs==15.0.1
dask-expr~=1.0.3
epochalyst==0.2
hydra-core==1.3.2
kaggle==1.6.6
numpy==1.26.4
pyarrow~=15.0.0
randomname==0.2.1
scikit-learn~=1.4.1
segmentation_models_pytorch==0.3.3
timm>=0.9.2
tqdm==4.66.2
wandb>=0.16.4

-e ./
